feat: Add routes for finding customer by contact and battery
The code changes include adding two new routes in `web.php` for finding a customer by contact and finding a battery.

In `QuickQuotation.php`, the `findCustomerByContact` function is added to search for customers based on a contact query. The `findBattery` function is also added to search for batteries based on an input query.

Recent user commits:
- Fixing bug in Sales Order
- Updating DataTables export function
- Updating DataTables price excel export
- Adding formatted date in Sales Order
- Fixing technician bug in Sales Order
- Fixing Sales Order subtotal bug
- Making email not mandatory in Customer
- Update QuickQuotation.php to include subtotal input in getPaymentPreview function
- Fix HTTP post duplication in QuickQuotation.php & remove btn marker on table sales order
- Fix typo in checkout page heading & sales order alamat