feat: Add distributor search functionality to QuickQuotation
This commit adds the `findDistributor` function to the `QuickQuotation.php` file. The function retrieves distributor data based on the provided input query. The retrieved data is then used to populate a combo box in the `step-1.blade.php` file. This feature enhances the user experience by allowing them to easily select a distributor from the available options.