feat: Add print technician report template functionality
This commit adds the functionality to print a technician report using a template. It includes the ability to select between printing from a default template or a template set in the database. The template is populated with data from the work order, including the work order number, battery information, and address. If no template is found in the database, an error message is displayed.