feat: Update print technician report template functionality
This commit updates the print technician report template functionality by allowing the selection of printing from a default template or a template set in the database. It also includes the addition of a QR code generated using the work order ID for quick access to the work order details. The QR code is displayed on the technician report template view.

Closes #123