refactor: Add battery code column to BatteryModel allForDataTables method
The commit modifies the `allForDataTables` method in the `BatteryModel` class to include the `battery_codes.code` column in the select and search columns. This change allows the battery codes to be displayed and searched in the DataTables view. Additionally, the method now joins the `battery_codes` table to retrieve the battery codes associated with each battery. The commit also includes a sorting parameter to order the results by the `batteries.updated_at` column in descending order.

Closes #789