feat: Update BatteryModel allForDataTables method to include battery code column
The commit modifies the `allForDataTables` method in the `BatteryModel` class to include the `battery_codes.code` column in the select and search columns. This change allows the battery codes to be displayed and searched in the DataTables view. Additionally, the method now joins the `battery_codes` table to retrieve the battery codes associated with each battery.