feat: Update composer.json and QuickQuotation index.blade.php
The commit updates the composer.json file to use a specific version of the setasign/fpdi dependency. It changes the version from "*" to "^2.6". This ensures that the project uses a compatible version of the library.

In addition, the commit modifies the QuickQuotation index.blade.php file. It adds two new variables, Latitude and Longitude, to the JavaScript code. These variables are used for address selection and validation. If either Latitude or Longitude is empty, an error message is displayed.

Closes #