Refactor: Update SalesOrderBatteryModel to include battery_price_retail field
Add the 'battery_price_retail' field to the SalesOrderBatteryModel class to store the retail price of the battery.

Refactor: Add source_id column to sales_orders table

Create a migration to add the 'source_id' column to the 'sales_orders' table with a default value of 0.

Refactor: Add source_platform column to sales_orders table

Create a migration to add the 'source_platform' column to the 'sales_orders' table with a default value of 'internal'.

Refactor: Update middleware permissions in routes/web.php

Add new routes and middleware permissions for the 'sales-online' functionality in the web.php file.

Refactor: Update SalesOrder controller to display source_platform badge

Update the SalesOrder controller to display a badge indicating the source platform of the sales order.